backend/requirements.txt
Django>=5.2
django-axes
djangorestframework
djangorestframework-simplejwt
django-environ
django-autocomplete-light
django-cors-headers
psycopg2
pillow
requests
regex
email-validator
pytz
python-decouple
gunicorn
ghostscript
PyPDF2
pdfkit
drf-spectacular